AK-50744 Update service-pan setInstance function (#346)
Update service-pan setInstance function to assume that API response
returns instance as ordered list.

With the ordered instances in response, there is no need to traverse the state
and API response twice to try to match instances. However, with the two hidden
credential variables, we still need to search existing TF state to locate them,
otherwise, TF will still generate a diff. Also, with hidden credentials, there
could be other potential issues if resource is modified outside of TF. That's
not what we could solve in this bug.